services/orders/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
pydantic-settings==2.5.2
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
nats-py==2.8.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
python-dotenv==1.0.1
